mcp-server/release-dashboard/pyproject.toml
[project]
name = "release-dashboard-mcp"
version = "1.0.0"
description = "MCP Server for EV2 Release Dashboard - queries Azure AppConfig deployment rollout status"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.0.0",
    "httpx>=0.27.0",
    "msal>=1.28.0",
    "pyyaml>=6.0",
]

[project.scripts]
release-dashboard-mcp = "release_dashboard_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/release_dashboard_mcp"]
